+++ server_src/runners/printerTicker.js
@@ -1,6 +1,30 @@
 const currentIssues = [];
+const octoprintLogs = [];
 
 class PrinterTicker {
+  static async addOctoPrintLog(printer, message, state, plugin) {
+    let id = null;
+    if (octoprintLogs.length === 0) {
+      //first issue
+      id = 0;
+    } else {
+      id = octoprintLogs[octoprintLogs.length - 1].id + 1;
+    }
+    const newLog = {
+      id: id,
+      date: new Date(),
+      message: message,
+      printerID: printer._id,
+      printer: printer.printerURL,
+      state: state,
+      pluginDisplay: plugin,
+    };
+    octoprintLogs.push(newLog);
+    if (octoprintLogs.length >= 10000) {
+      octoprintLogs.shift();
+    }
+  }
+
   static async addIssue(date, printer, message, state, printerID) {
     let id = null;
     if (currentIssues.length === 0) {
@@ -27,6 +51,9 @@ class PrinterTicker {
       return o.id == id;
     });
     currentIssues.splice(index, 1);
+  }
+  static async returnOctoPrintLogs() {
+    return octoprintLogs;
   }
   static async returnIssue() {
     return currentIssues;